Add AnimationStyle file
